travel-agent/backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
greenlet==3.1.1

# Settings
pydantic-settings==2.6.1
pydantic==2.10.1

# AI
anthropic==0.40.0

# Auth
python-jose[cryptography]==3.3.0
httpx==0.28.0

# AWS (optional in dev — injected via env)
boto3==1.35.84

# Google APIs (optional in dev)
google-api-python-client==2.154.0
google-auth==2.37.0

# Dev/test
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.0
